Move dataset.py into its own file and add CustomEncoder
